Refactor temper-spec for guideline compliance
