+++ crates/temper-spec/src/automaton/parser.rs
@@ -230,592 +230,5 @@ fn validate(automaton: &Automaton) -> Result<(), AutomatonParseError> {
 }
 
 #[cfg(test)]
-mod tests {
-    use super::*;
-
-    const ORDER_IOA: &str = include_str!("../../../../test-fixtures/specs/order.ioa.toml");
-
-    #[test]
-    fn test_parse_order_automaton() {
-        let automaton = parse_automaton(ORDER_IOA).expect("should parse");
-        assert_eq!(automaton.automaton.name, "Order");
-        assert_eq!(automaton.automaton.initial, "Draft");
-        assert_eq!(automaton.automaton.states.len(), 10);
-        assert!(automaton.automaton.states.contains(&"Draft".to_string()));
-        assert!(automaton.automaton.states.contains(&"Shipped".to_string()));
-    }
-
-    #[test]
-    fn test_actions_parsed() {
-        let automaton = parse_automaton(ORDER_IOA).unwrap();
-        let names: Vec<&str> = automaton.actions.iter().map(|a| a.name.as_str()).collect();
-        assert!(names.contains(&"AddItem"), "got: {names:?}");
-        assert!(names.contains(&"SubmitOrder"));
-        assert!(names.contains(&"CancelOrder"));
-        assert!(names.contains(&"ConfirmOrder"));
-    }
-
-    #[test]
-    fn test_submit_order_has_guard() {
-        let automaton = parse_automaton(ORDER_IOA).unwrap();
-        let submit = automaton
-            .actions
-            .iter()
-            .find(|a| a.name == "SubmitOrder")
-            .unwrap();
-        assert_eq!(submit.from, vec!["Draft"]);
-        assert_eq!(submit.to, Some("Submitted".to_string()));
-        assert!(!submit.guard.is_empty(), "SubmitOrder should have a guard");
-    }
-
-    #[test]
-    fn test_cancel_from_multiple_states() {
-        let automaton = parse_automaton(ORDER_IOA).unwrap();
-        let cancel = automaton
-            .actions
-            .iter()
-            .find(|a| a.name == "CancelOrder")
-            .unwrap();
-        assert_eq!(cancel.from.len(), 3);
-        assert!(cancel.from.contains(&"Draft".to_string()));
-        assert!(cancel.from.contains(&"Submitted".to_string()));
-        assert!(cancel.from.contains(&"Confirmed".to_string()));
-    }
-
-    #[test]
-    fn test_invariants_parsed() {
-        let automaton = parse_automaton(ORDER_IOA).unwrap();
-        assert!(!automaton.invariants.is_empty());
-        let names: Vec<&str> = automaton
-            .invariants
-            .iter()
-            .map(|i| i.name.as_str())
-            .collect();
-        assert!(names.contains(&"SubmitRequiresItems"), "got: {names:?}");
-    }
-
-    #[test]
-    fn test_convert_to_state_machine() {
-        let automaton = parse_automaton(ORDER_IOA).unwrap();
-        let sm = to_state_machine(&automaton);
-        assert_eq!(sm.module_name, "Order");
-        assert_eq!(sm.states.len(), 10);
-        assert!(!sm.transitions.is_empty());
-        assert!(!sm.invariants.is_empty());
-
-        let submit = sm
-            .transitions
-            .iter()
-            .find(|t| t.name == "SubmitOrder")
-            .unwrap();
-        assert_eq!(submit.from_states, vec!["Draft"]);
-        assert_eq!(submit.to_state, Some("Submitted".to_string()));
-    }
-
-    #[test]
-    fn test_invalid_initial_state_rejected() {
-        let toml = r#"
-[automaton]
-name = "Bad"
-states = ["A", "B"]
-initial = "C"
-"#;
-        let result = parse_automaton(toml);
-        assert!(result.is_err());
-    }
-
-    #[test]
-    fn test_invalid_from_state_rejected() {
-        let toml = r#"
-[automaton]
-name = "Bad"
-states = ["A", "B"]
-initial = "A"
-
-[[action]]
-name = "Go"
-from = ["Z"]
-to = "B"
-"#;
-        let result = parse_automaton(toml);
-        assert!(result.is_err());
-    }
-
-    #[test]
-    fn test_integration_section_parsed() {
-        let toml = r#"
-[automaton]
-name = "Order"
-states = ["Draft", "Submitted"]
-initial = "Draft"
-
-[[action]]
-name = "SubmitOrder"
-from = ["Draft"]
-to = "Submitted"
-
-[[integration]]
-name = "notify_fulfillment"
-trigger = "SubmitOrder"
-type = "webhook"
-
-[[integration]]
-name = "charge_payment"
-trigger = "ConfirmOrder"
-type = "webhook"
-"#;
-        let automaton = parse_automaton(toml).expect("should parse");
-        assert_eq!(automaton.integrations.len(), 2);
-        assert_eq!(automaton.integrations[0].name, "notify_fulfillment");
-        assert_eq!(automaton.integrations[0].trigger, "SubmitOrder");
-        assert_eq!(automaton.integrations[0].integration_type, "webhook");
-        assert_eq!(automaton.integrations[1].name, "charge_payment");
-    }
-
-    #[test]
-    fn test_integration_default_type() {
-        let toml = r#"
-[automaton]
-name = "Order"
-states = ["Draft", "Submitted"]
-initial = "Draft"
-
-[[integration]]
-name = "notify"
-trigger = "SubmitOrder"
-"#;
-        let automaton = parse_automaton(toml).expect("should parse");
-        assert_eq!(automaton.integrations.len(), 1);
-        assert_eq!(automaton.integrations[0].integration_type, "webhook");
-    }
-
-    #[test]
-    fn test_no_integrations_defaults_empty() {
-        let automaton = parse_automaton(ORDER_IOA).expect("should parse");
-        assert!(automaton.integrations.is_empty() || !automaton.integrations.is_empty());
-    }
-
-    #[test]
-    fn test_trigger_effect_parsed() {
-        let toml = r#"
-[automaton]
-name = "Order"
-states = ["Submitted", "ChargePending", "Confirmed", "PaymentFailed"]
-initial = "Submitted"
-
-[[action]]
-name = "ChargePayment"
-from = ["Submitted"]
-to = "ChargePending"
-effect = "trigger charge_payment"
-
-[[action]]
-name = "ChargeSucceeded"
-kind = "input"
-from = ["ChargePending"]
-to = "Confirmed"
-
-[[action]]
-name = "ChargeFailed"
-kind = "input"
-from = ["ChargePending"]
-to = "PaymentFailed"
-"#;
-        let automaton = parse_automaton(toml).expect("should parse");
-        let charge = automaton
-            .actions
-            .iter()
-            .find(|a| a.name == "ChargePayment")
-            .unwrap();
-        assert_eq!(charge.effect.len(), 1);
-        match &charge.effect[0] {
-            Effect::Trigger { name } => assert_eq!(name, "charge_payment"),
-            other => panic!("expected Trigger effect, got: {other:?}"),
-        }
-    }
-
-    #[test]
-    fn test_wasm_integration_parsed() {
-        let toml = r#"
-[automaton]
-name = "Order"
-states = ["Submitted", "ChargePending", "Confirmed", "PaymentFailed"]
-initial = "Submitted"
-
-[[action]]
-name = "ChargePayment"
-from = ["Submitted"]
-to = "ChargePending"
-effect = "trigger charge_payment"
-
-[[action]]
-name = "ChargeSucceeded"
-kind = "input"
-from = ["ChargePending"]
-to = "Confirmed"
-
-[[action]]
-name = "ChargeFailed"
-kind = "input"
-from = ["ChargePending"]
-to = "PaymentFailed"
-
-[[integration]]
-name = "charge_payment"
-trigger = "charge_payment"
-type = "wasm"
-module = "stripe_charge"
-on_success = "ChargeSucceeded"
-on_failure = "ChargeFailed"
-"#;
-        let automaton = parse_automaton(toml).expect("should parse");
-        assert_eq!(automaton.integrations.len(), 1);
-        let ig = &automaton.integrations[0];
-        assert_eq!(ig.name, "charge_payment");
-        assert_eq!(ig.integration_type, "wasm");
-        assert_eq!(ig.module.as_deref(), Some("stripe_charge"));
-        assert_eq!(ig.on_success.as_deref(), Some("ChargeSucceeded"));
-        assert_eq!(ig.on_failure.as_deref(), Some("ChargeFailed"));
-    }
-
-    #[test]
-    fn test_wasm_integration_missing_module_rejected() {
-        let toml = r#"
-[automaton]
-name = "Order"
-states = ["Submitted", "ChargePending"]
-initial = "Submitted"
-
-[[action]]
-name = "ChargePayment"
-from = ["Submitted"]
-to = "ChargePending"
-
-[[integration]]
-name = "charge_payment"
-trigger = "charge_payment"
-type = "wasm"
-"#;
-        let result = parse_automaton(toml);
-        assert!(result.is_err());
-        let err = result.unwrap_err().to_string();
-        assert!(err.contains("missing 'module'"), "got: {err}");
-    }
-
-    #[test]
-    fn test_wasm_integration_unknown_callback_rejected() {
-        let toml = r#"
-[automaton]
-name = "Order"
-states = ["Submitted", "ChargePending", "Confirmed"]
-initial = "Submitted"
-
-[[action]]
-name = "ChargePayment"
-from = ["Submitted"]
-to = "ChargePending"
-
-[[integration]]
-name = "charge_payment"
-trigger = "charge_payment"
-type = "wasm"
-module = "stripe_charge"
-on_success = "NonExistentAction"
-"#;
-        let result = parse_automaton(toml);
-        assert!(result.is_err());
-        let err = result.unwrap_err().to_string();
-        assert!(
-            err.contains("NonExistentAction"),
-            "should mention missing action, got: {err}"
-        );
-    }
-
-    #[test]
-    fn test_valid_state_var_types_accepted() {
-        let spec = r#"
-[automaton]
-name = "Task"
-states = ["Open", "Done"]
-initial = "Open"
-
-[[state]]
-name = "is_done"
-type = "bool"
-initial = "false"
-
-[[state]]
-name = "attempt_count"
-type = "counter"
-initial = "0"
-
-[[action]]
-name = "Complete"
-kind = "input"
-from = ["Open"]
-to = "Done"
-effect = "set is_done true"
-"#;
-        let result = parse_automaton(spec);
-        assert!(
-            result.is_ok(),
-            "bool and counter types should be accepted: {:?}",
-            result.err()
-        );
-    }
-
-    #[test]
-    fn test_extended_guard_syntax_parsed() {
-        let spec = r#"
-[automaton]
-name = "Ticket"
-states = ["Open", "Queued", "Closed"]
-initial = "Open"
-
-[[action]]
-name = "Queue"
-from = ["Open"]
-to = "Queued"
-guard = "max retries 3"
-
-[[action]]
-name = "Escalate"
-from = ["Queued"]
-to = "Queued"
-guard = "list_contains labels urgent"
-
-[[action]]
-name = "Close"
-from = ["Queued"]
-to = "Closed"
-guard = "list_length_min labels 1"
-"#;
-
-        let automaton = parse_automaton(spec).expect("extended guard forms should parse");
-        let queue = automaton
-            .actions
-            .iter()
-            .find(|a| a.name == "Queue")
-            .unwrap();
-        assert!(matches!(
-            queue.guard.as_slice(),
-            [Guard::MaxCount { var, max }] if var == "retries" && *max == 3
-        ));
-
-        let escalate = automaton
-            .actions
-            .iter()
-            .find(|a| a.name == "Escalate")
-            .unwrap();
-        assert!(matches!(
-            escalate.guard.as_slice(),
-            [Guard::ListContains { var, value }] if var == "labels" && value == "urgent"
-        ));
-
-        let close = automaton
-            .actions
-            .iter()
-            .find(|a| a.name == "Close")
-            .unwrap();
-        assert!(matches!(
-            close.guard.as_slice(),
-            [Guard::ListLengthMin { var, min }] if var == "labels" && *min == 1
-        ));
-    }
-
-    #[test]
-    fn test_integration_config_captures_unknown_keys() {
-        let toml = r#"
-[automaton]
-name = "Weather"
-states = ["Idle", "Fetching", "Ready", "Failed"]
-initial = "Idle"
-
-[[action]]
-name = "FetchWeather"
-from = ["Idle"]
-to = "Fetching"
-effect = "trigger fetch_weather"
-
-[[action]]
-name = "FetchSucceeded"
-kind = "input"
-from = ["Fetching"]
-to = "Ready"
-
-[[action]]
-name = "FetchFailed"
-kind = "input"
-from = ["Fetching"]
-to = "Failed"
-
-[[integration]]
-name = "fetch_weather"
-trigger = "fetch_weather"
-type = "wasm"
-module = "http_fetch"
-on_success = "FetchSucceeded"
-on_failure = "FetchFailed"
-url = "https://api.open-meteo.com/v1/forecast"
-method = "GET"
-"#;
-        let automaton = parse_automaton(toml).expect("should parse");
-        assert_eq!(automaton.integrations.len(), 1);
-        let ig = &automaton.integrations[0];
-        assert_eq!(ig.name, "fetch_weather");
-        assert_eq!(ig.integration_type, "wasm");
-        assert_eq!(ig.module.as_deref(), Some("http_fetch"));
-        assert_eq!(
-            ig.config.get("url").map(String::as_str),
-            Some("https://api.open-meteo.com/v1/forecast")
-        );
-        assert_eq!(ig.config.get("method").map(String::as_str), Some("GET"));
-        // Known keys should NOT be in config
-        assert!(!ig.config.contains_key("name"));
-        assert!(!ig.config.contains_key("trigger"));
-        assert!(!ig.config.contains_key("type"));
-        assert!(!ig.config.contains_key("module"));
-    }
-
-    #[test]
-    fn test_invalid_guard_number_rejected() {
-        let spec = r#"
-[automaton]
-name = "Order"
-states = ["Draft", "Submitted"]
-initial = "Draft"
-
-[[action]]
-name = "SubmitOrder"
-from = ["Draft"]
-to = "Submitted"
-guard = "items > nope"
-"#;
-
-        let err = parse_automaton(spec).expect_err("invalid numeric guard should fail");
-        assert!(err.to_string().contains("right side must be an integer"));
-    }
-
-    #[test]
-    fn test_parse_schedule_effect() {
-        let spec = r#"
-[automaton]
-name = "OAuthToken"
-states = ["Active", "Refreshing", "Expired"]
-initial = "Active"
-
-[[action]]
-name = "Activate"
-from = ["Refreshing"]
-to = "Active"
-effect = [{ type = "schedule", action = "Refresh", delay_seconds = 2700 }]
-"#;
-
-        let automaton = parse_automaton(spec).expect("should parse schedule effect");
-        let activate = automaton
-            .actions
-            .iter()
-            .find(|a| a.name == "Activate")
-            .unwrap();
-        assert_eq!(activate.effect.len(), 1);
-        match &activate.effect[0] {
-            Effect::Schedule {
-                action,
-                delay_seconds,
-            } => {
-                assert_eq!(action, "Refresh");
-                assert_eq!(*delay_seconds, 2700);
-            }
-            other => panic!("expected Schedule, got: {:?}", other),
-        }
-    }
-
-    #[test]
-    fn test_unknown_inline_effect_type_rejected() {
-        let spec = r#"
-[automaton]
-name = "Broken"
-states = ["Draft", "Done"]
-initial = "Draft"
-
-[[action]]
-name = "Complete"
-from = ["Draft"]
-to = "Done"
-effect = [{ type = "mystery_effect", value = "x" }]
-"#;
-        let err = parse_automaton(spec).expect_err("unknown inline effect type should fail");
-        assert!(
-            err.to_string()
-                .contains("unsupported effect type 'mystery_effect'")
-        );
-    }
-
-    #[test]
-    fn test_legacy_inline_effect_aliases_supported() {
-        let spec = r#"
-[automaton]
-name = "Plan"
-states = ["Active"]
-initial = "Active"
-
-[[action]]
-name = "AddTask"
-from = ["Active"]
-effect = [
-  { type = "spawn_entity", entity_type = "Task", entity_id_source = "{uuid}", initial_action = "Create" },
-  { type = "emit_event", event = "TaskAdded" }
-]
-"#;
-        let automaton = parse_automaton(spec).expect("legacy aliases should parse");
-        let add_task = automaton
-            .actions
-            .iter()
-            .find(|a| a.name == "AddTask")
-            .expect("AddTask action should exist");
-        assert!(matches!(
-            add_task.effect.first(),
-            Some(Effect::Spawn { .. })
-        ));
-        assert!(matches!(add_task.effect.get(1), Some(Effect::Emit { .. })));
-    }
-
-    #[test]
-    fn test_agent_trigger_parsed() {
-        let spec = r#"
-[automaton]
-name = "Project"
-states = ["Draft", "Ready"]
-initial = "Draft"
-
-[[action]]
-name = "MarkReady"
-from = ["Draft"]
-to = "Ready"
-
-[[agent_trigger]]
-name = "test_on_ready"
-on_action = "MarkReady"
-to_state = "Ready"
-agent_role = "tester"
-agent_goal = "Run integration tests"
-agent_type_id = "tester-type-1"
-"#;
-        let automaton = parse_automaton(spec).expect("agent_trigger should parse");
-        assert_eq!(automaton.agent_triggers.len(), 1);
-        let trigger = &automaton.agent_triggers[0];
-        assert_eq!(trigger.name, "test_on_ready");
-        assert_eq!(trigger.on_action, "MarkReady");
-        assert_eq!(trigger.to_state, Some("Ready".to_string()));
-        assert_eq!(trigger.agent_role, "tester");
-        assert_eq!(trigger.agent_goal, "Run integration tests");
-        assert_eq!(trigger.agent_type_id, Some("tester-type-1".to_string()));
-        assert!(trigger.agent_model.is_none());
-    }
-
-    #[test]
-    fn test_agent_trigger_defaults_empty() {
-        let automaton = parse_automaton(ORDER_IOA).expect("should parse");
-        assert!(automaton.agent_triggers.is_empty());
-    }
-}
+#[path = "parser_test.rs"]
+mod tests;
